pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hearth"
version = "0.1.0"
description = "Personal Cognitive Operating System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-only"}
authors = [
    {name = "HEARTH System", email = "security@hearth.local"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "chromadb>=0.4.0",
    "ollama>=0.1.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "pdfplumber>=0.10.0",
    "Pillow>=10.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.urls]
"Homepage" = "https://github.com/your-org/hearth"
"Bug Tracker" = "https://github.com/your-org/hearth/issues"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E", "W", "F",  # Default pycodestyle errors
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = ["E501", "B008"]